accounted for pass move

+++ NNet.py
@@ -52,11 +52,11 @@ class NNetWrapper():
             v_losses = AverageMeter()
             end = time.time()
 
-            bar = Bar('Processing', max=len(examples)/args.batch_size)
+            bar = Bar('Processing', max=int(len(examples)/args.batch_size))
             batch_idx = 0
 
-            while batch_idx < len(examples)/args.batch_size:
-                sample_ids = np.random.randint(len(examples), size=args.batch_size)                
+            while batch_idx < int(len(examples)/args.batch_size):
+                sample_ids = np.random.randint(len(examples), size=args.batch_size)
                 boards, pis, vs = list(zip(*[examples[i] for i in sample_ids]))
                 boards = torch.FloatTensor(np.array(boards).astype(np.float64))
                 target_pis = torch.FloatTensor(np.array(pis))
